# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "HouseFinder API",
    "description": "Lista degli endpoints dell'applicazione HouseFinder e informazioni sul loro utilizzo."
  },
  "host": "localhost:3000",
  "basePath": "/",
  "security": [
    {
      "jwt": []
    }
  ],
  "securityDefinitions": {
    "jwt": {
      "name": "x-access-token",
      "type": "apiKey",
      "in": "header"
    }
  },
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/annuncio/list": {
      "get": {
        "summary": "Lista annunci",
        "description": "Restituisce la lista degli annunci della pagina principale",
        "operationId": "get_annunci",
        "tags": [
          "annuncio"
        ],
        "responses": {
          "200": {
            "description": "La richiesta è stata eseguita con successo.",
            "schema": {
              "properties": {
                "annunci": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/annuncio"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Annunci non trovati."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/annuncio/{id}": {
      "get": {
        "summary": "Visualizza annuncio",
        "description": "Restituisce le informazioni di un annuncio",
        "operationId": "get_annuncio",
        "tags": [
          "annuncio"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id annuncio",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Annuncio ottenuto correttamente.",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/annuncio"
              }
            }
          },
          "404": {
            "description": "Annuncio non trovato."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      },
      "patch": {
        "summary": "Modifica annuncio",
        "description": "Modifica un annuncio pubblicato dall'utente",
        "operationId": "modifica_annuncio",
        "tags": [
          "annuncio"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id dell'annuncio da modificare",
            "required": true,
            "type": "string"
          },
          {
            "name": "Dati annuncio",
            "in": "body",
            "description": "Dati dell'annuncio da modificare",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pubblica_annuncio"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Modifica dell'annuncio riuscita."
          },
          "404": {
            "description": "Annuncio non trovato."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      },
      "delete": {
        "summary": "Elimina annuncio",
        "description": "Elimina un annuncio pubblicato dall'utente",
        "operationId": "elimina_annuncio",
        "tags": [
          "annuncio"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id dell'annuncio da eliminare",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Annuncio eliminato con successo."
          },
          "404": {
            "description": "Annuncio non trovato."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/annuncio": {
      "post": {
        "summary": "Pubblica annuncio",
        "description": "Pubblica un annuncio",
        "operationId": "pubblica_annuncio",
        "tags": [
          "annuncio"
        ],
        "parameters": [
          {
            "name": "annuncio",
            "in": "body",
            "description": "Annuncio da pubblicare",
            "schema": {
              "$ref": "#/definitions/pubblica_annuncio"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Annuncio creato con successo."
          },
          "400": {
            "description": "Alcuni parametri sono assenti."
          },
          "401": {
            "description": "Utente non autorizzato."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/annuncio/salva/{id}": {
      "get": {
        "summary": "Salva annuncio",
        "description": "Salva un annuncio nella lista degli annunci salvati",
        "operationId": "salva_annuncio",
        "tags": [
          "annuncio"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id annuncio",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Annuncio salvato con successo."
          },
          "404": {
            "description": "Annuncio non trovato."
          },
          "409": {
            "description": "Annuncio già presente negli annunci salvati."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      },
      "delete": {
        "summary": "Rimuovi annuncio salvato",
        "description": "Rimuove un annuncio dalla lista degli annunci salvati",
        "operationId": "rimuovi_annuncio_salvato",
        "tags": [
          "annuncio"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id annuncio",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Annuncio rimosso con successo."
          },
          "404": {
            "description": "Annuncio non trovato."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/annuncio/pagamento": {
      "post": {
        "summary": "Pagamento",
        "description": "Effettua il pagamento della vetrina di un annuncio",
        "operationId": "pagamento",
        "tags": [
          "annuncio"
        ],
        "parameters": [
          {
            "name": "annuncio_id",
            "in": "body",
            "description": "Annuncio da pagare",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Pagamento effettuato."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/ricerca": {
      "post": {
        "summary": "Ricerca annunci",
        "description": "Ricerca annunci nel sistema",
        "operationId": "ricerca_annunci",
        "tags": [
          "ricerca"
        ],
        "parameters": [
          {
            "name": "filtri",
            "in": "body",
            "description": "Filtri ricerca",
            "schema": {
              "$ref": "#/definitions/ricerca_filtri"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Elenco degli annunci ottenuto correttamente.",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/annuncio"
              }
            }
          },
          "404": {
            "description": "Annunci non trovati."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/ricerca/salva": {
      "post": {
        "summary": "Salva ricerca",
        "description": "Salva una ricerca",
        "operationId": "salva_ricerca",
        "tags": [
          "ricerca"
        ],
        "parameters": [
          {
            "name": "parametri",
            "in": "body",
            "description": "Parametri ricerca",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ricerca_input"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ricerca salvata con successo."
          },
          "404": {
            "description": "Utente non trovato."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/ricerca/salva/{id}": {
      "delete": {
        "summary": "Rimuovi ricerca salvata",
        "description": "Rimuove una ricerca dalla lista delle ricerche salvate",
        "operationId": "rimuovi_ricerca_salvata",
        "tags": [
          "ricerca"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id ricerca",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Ricerca rimossa con successo."
          },
          "404": {
            "description": "Utente non trovato."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "annuncio": {
      "type": "object",
      "properties": {
        "creatore": {
          "type": "string"
        },
        "foto": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "superficie_tot": {
          "type": "number"
        },
        "numero_bagni": {
          "type": "number"
        },
        "numero_locali": {
          "type": "number"
        },
        "locali": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/locale"
          }
        },
        "prezzo": {
          "type": "number"
        },
        "classe_energetica": {
          "type": "string"
        },
        "indirizzo": {
          "type": "string"
        },
        "arredato": {
          "type": "boolean"
        },
        "vetrina": {
          "properties": {
            "data_inizio": {
              "$ref": "#/definitions/date"
            },
            "data_fine": {
              "$ref": "#/definitions/date"
            }
          }
        }
      }
    },
    "locale": {
      "type": "object",
      "properties": {
        "nome_locale": {
          "type": "string"
        },
        "superficie": {
          "type": "number"
        },
        "descrizione": {
          "type": "string"
        },
        "arredamento": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
...
    "filtro": {
      "type": "object",
      "properties": {
        "min": {
          "type": "number"
        },
        "max": {
          "type": "number"
        }
      }
    },
...
    "pubblica_annuncio": {
      "type": "object",
      "properties": {
        "userId": {
          "type": "string"
        },
        "foto": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "superficie_tot": {
          "type": "number"
        },
        "numero_bagni": {
          "type": "number"
        },
        "numero_locali": {
          "type": "number"
        },
        "locali": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/locale"
          }
        },
        "prezzo": {
          "type": "number"
        },
        "classe_energetica": {
          "type": "string"
        },
        "indirizzo": {
          "type": "string"
        },
        "arredato": {
          "type": "boolean"
        },
        "durata_vetrina": {
          "type": "number"
        }
      }
    },
    "ricerca_input": {
      "type": "object",
      "properties": {
        "superficie_tot": {
          "type": "number"
        },
        "numero_bagni": {
          "type": "number"
        },
        "numero_locali": {
          "type": "number"
        },
        "prezzo": {
          "type": "number"
        },
        "classe_energetica": {
          "type": "string"
        },
        "indirizzo": {
          "type": "string"
        },
        "arredato": {
          "type": "boolean"
        }
      }
    },
...
    "ricerca_filtri": {
      "type": "object",
      "properties": {
        "superficie_tot": {
          "$ref": "#/definitions/filtro"
        },
        "numero_bagni": {
          "$ref": "#/definitions/filtro"
        },
        "numero_locali": {
          "$ref": "#/definitions/filtro"
        },
        "prezzo": {
          "$ref": "#/definitions/filtro"
        },
        "classe_energetica": {
          "type": "string"
        },
        "indirizzo": {
          "type": "string"
        },
        "arredato": {
          "type": "boolean"
        }
      }
    },
    "date": {
      "type": "string"
    }
  }
}